Update README.md
This repository contains all of the programs written and executed to complete this third year project. The code is split into two branches. The first containing the Python code used to build and test the proposed model and to send the testing dataset to the edge microcontroller via serial communication. The second folder contains the C++ code used to both receive the testing data set from the PC and invoke the edge-deployed TensorFlow Lite model as shown.